i18n: move out from httperrors pkg
