+++ pkg/i18n/doc.go
@@ -1,0 +1,1 @@
+package i18n // import "yunion.io/x/onecloud/pkg/i18n"
