+++ pkg/i18n/print.go
@@ -1,4 +1,4 @@
-package httperrors
+package i18n
 
 import (
 	"golang.org/x/text/language"
